incorporating menu-ui into struct
